Construct CameraProjection from 3x4 matrix
In order to achieve more realistic projections I suppose to allow
the construction of CameraProjections from 3x4 matrices.

To do so, the 3x4 matrix is decomposed using a RQ-decomposition
into the intrinsics (K) and extrinsics (R, t). Then, the existing
classmethod _from_rtk()_ is reused.

Resources which were useful to comprehend the decomposition:

- https://ksimek.github.io/2012/08/14/decompose
- Siemens internal decomposition software